requirements.txt

streamlit==1.43.0
beautifulsoup4==4.12.3
requests==2.31.0
groq==0.15.0
python-dotenv==1.0.1
